refactored snippet location
