+++ compute/cloud-client/src/main/java/compute/windows/osimage/CreateWindowsOsImage.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package compute.windowsosimage;
+package compute.windows.osimage;
 
 // [START compute_windows_image_create]
