+++ compute/cloud-client/src/test/java/compute/windows/osimage/WindowsOsImageIT.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package compute.windowsosimage;
+package compute.windows.osimage;
 
 import static com.google.common.truth.Truth.assertThat;
 import static com.google.common.truth.Truth.assertWithMessage;
